Split flask_adapter.py to meet line limit
- Split flask_adapter.py (309→84 lines) by extracting request handlers
- Created request_handlers.py (175 lines) for request handling logic
- Updated flask_adapter.py to use RequestHandlers for delegation
- Progress: 3 modules still exceed 300-line limit, continuing refactoring
